app/requirements.txt
pandas==1.3.3
matplotlib==3.3.4
psycopg2==2.8.6
Flask==2.0.1
flask_apscheduler==1.12.2
protobuf==3.15.5
seaborn==0.11.1
slackclient==2.9.3
openpyxl==3.0.7
gunicorn==20.1.0
requests==2.25.1
redash-toolbelt==0.1.4
jira==3.0.1
redis==3.5.3
pygsheets==2.0.5